BOTV2-93
correction de chargement de la page

la chargement du page reste trop lent, à cause du loading requetes dans les différents widget

le changement c'est de séparer chaque widget dans une page séparé, pour alléger le chargement .
